MemoriDataProvider moved from tests into the main component.
