+++ src/SenseNet.Security.TestConsole/Program.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using SenseNet.Security.Data;
 
 namespace SenseNet.Security.TestConsole
 {
