+++ src/SenseNet.Security.Tests.Concurrency/Program.cs
@@ -8,6 +8,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using SenseNet.Security.Data;
 
 namespace SenseNet.Security.Tests.Concurrency
 {
