+++ src/SenseNet.Security.Tests/Accessors.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SenseNet.Security;
+using SenseNet.Security.Data;
 
 namespace SenseNet.Security.Tests
 {
