+++ src/SenseNet.Security.Tests/BigSystemStartTest.cs
@@ -9,6 +9,7 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using SenseNet.Security.Data;
 
 namespace SenseNet.Security.Tests
 {
