+++ src/SenseNet.Security.Tests/MembershipTests.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using SenseNet.Security.Data;
 
 namespace SenseNet.Security.Tests
 {
