+++ src/SenseNet.Security.Tests/SystemStartTests.cs
@@ -5,6 +5,7 @@ using System.Text;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Diagnostics;
 using SenseNet.Security;
+using SenseNet.Security.Data;
 using SenseNet.Security.Messaging;
 using SenseNet.Security.Tests.TestPortal;
